[All] Rework encryption type selection & AES User Key handling
Instead of changing a global variable, AES User Key is now passed to WzFile, WzDirectory, WzBinary Reader & Writer like WzIv is. This removes the need to manage the UserKey when Custom encryption is and isn't being used.

Moved handling of the dropdowns for encryption to HaSharedLibrary
Added Custom Encryption support for HaCreator
Added "auto detect" to HaRepacker
Auto Detect now attempts the Custom encryption when detecting

This also fixes a bug when saving to Custom encryption that it would not revert back to the default key unless you changed encryptions then went back to custom.