[CanvasHelper] Initial version
This is a temp workaround for Canvas files. Its intended use is to remove canvas outlinks.

If the Canvas file does not include a specific .img, it will not import it. So this tool should not be used to attempt to create a fully working wz for lower versions. It is currently only for reducing the effort in fixing imgs that reference canvas entries.

Tool is rather simple. It loads all the canvas wz files & 'original' wz files, loops all the imgs in the canvas wzs, and pulls the original wz data to merge the 2 together, saving into a new canvas file.

Saves back to the multiple wz format to avoid the wz file being too large.